feat: Display shop list
Displayed shop list on the main screen with 100 items.
Used a `ScrollView` with a `LinearLayout` to
 show the items.
Enabled changing the state of an item by long pressing on it.
Removed the `RecyclerView` from the layout.
Updated `ShopListRepositoryImpl` to use a `sortedSetOf` for storing items.
Changed `deleteShopItem` to take a `ShopItem` object instead
 of an ID.