feat: Add landscape orientation support
Added a new layout for landscape orientation.
In landscape orientation, the app uses a two-pane layout with a list of
 items on the left and a details view on the right.
In portrait orientation, the app uses a single-pane layout with a list of items and a separate activity for the details view.
Removed unused code from ShopItemActivity.
Added logs to ShopItemFragment.